+++ export.js
@@ -1,6 +1,7 @@
 module.exports = {
     MudletMapReader : {
-        read: require("./map-reader"),
+        read: require("./map-operations").readMap,
+        write: require("./map-operations").writeMap,
         export: require("./reader-export"),
         exportJson: require('./json-export')
     }